seperate interfaces and query key into seperate file
