+++ src/api/types.ts
@@ -1,0 +1,11 @@
+export interface CatTributes {
+  id: string;
+  url: string;
+  width: number;
+  height: number;
+}
+
+export interface VoteProps {
+  image_id: string;
+  value: 1 | -1;
+}
